requirements.txt
django==4.0.6
djangorestframework==3.14.0
reportlab==3.6.12
python-dateutil==2.8.2
xhtml2pdf==0.2.9